Add convenient test function
This reads the first word from a sector, increments it, erases the
sector, writes the incremented word, and reads the sector and validates
the write was successful. You can examine the memory using
the TrueStudio memory analyzer or ST-Link Utility to verify
that the incremented value is correct, despite not being able
to reprogram the board.