Improve UART IRQ handler semantics
Previously, the UART IRQ handler would always try and abort a pending transaction if the idle flag was set, even if idle interrupts were not enabled. This change uses the already-existing `__HAL_UART_GET_IT_SOURCE()` macro to determine if the interrupt was actually caused by an idle interrupt. It also moves the idle handling to run before the HAL interrupt handler, since aborting a pending receive operation causes changes to the UART device state.

Finally, this change ensures the idle interrupt is initially disabled, since that's what the rest of the RX implementation assumes.